backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
mssql-django==1.6
pyodbc==5.2.0